requirements.txt
mysqlclient==2.1.0
numpy==1.21.4
pandas==1.3.4
psycopg2-binary==2.9.2
pytz==2021.3